+++ regexp/ChangeLog.md
@@ -1,9 +1,11 @@
-September 2017
+September 2017: 0.2.0.0
    - Convert to stack (nightly)
    - Use GHC.Records instead of HasField (GHC 8.2 required)
    - Broke dictionary code into OccDict.hs
    - Explicit export lists
-
+   - Used DList to avoid (++) in marks
+   - More haddocks
+   - Renamed R to RE (for clarity)
 
 January 2071: 0.1.0.0
    Initial version